chore(yarn): switch to node-modules linker
Using Vite with PnP linker is problematic. We are switching back to node-modules linker also because it didn't bring us the benefits we originally expected, since this repository is only one in our organization that is using PnP linker.

BREAKING CHANGE:

This PR takes advantage of changing yarn.lock format and refresh its whole content. This will lead to large number of changes in underlying package versions.