chore(release): remove necessity for bump commits
To comply with organization policies we are no longer able to push bump commits to the default branch. Per recommendation of the 'semantic-release' we opt-out from updating version in `package.json` (version is still updated in released packages) and providing changelog file (detailed changelogs are available in Github Releases page).