feat(DatatableV2): center cell content vertically
The decision to top align was mostly to account for use cases where we are displaying a lot of content in a cell. In retrospect, this was optimising for undesirable use cases.

Aligning content of varying sizes with top aligned cells is difficult and leads to visual artefacts in the most common use case when rows are not tall. By aligning to center we may create some issues in already problematic layouts, but will make working with most use cases easier.

Closes UXD-1627