feat(DatatableV2): add option to disable selection toolbar
For special cases we need to be able to disable the selection toolbar with batch actions while the selection itself stays enabled. In this setup is necessary to handle the selection state in managed way otherwise there is no other option how to access the state and act upon it.

Closes UXD-1647