test(a11y): add automated a11y testing
This PR sets up automated accessibility testing for all components in the Storybook. Those test are not yet required for successful deployment of the Design System to not block all development. We need to first review currently failing accessibility tests and either fix the issues or disable tests if they are false positive.

Closes UXD-1229